+++ app.config.js
@@ -178,18 +178,22 @@ export default {
 
 export const KnownDidRegistries = [
   {
+    type: 'dcc-legacy',
     name: 'DCC Pilot Registry',
     url: 'https://digitalcredentials.github.io/issuer-registry/registry.json'
   },
   {
+    type: 'dcc-legacy',
     name: 'DCC Sandbox Registry',
     url: 'https://digitalcredentials.github.io/sandbox-registry/registry.json'
   },
   {
+    type: 'dcc-legacy',
     name: 'DCC Community Registry',
     url: 'https://digitalcredentials.github.io/community-registry/registry.json'
   },
   {
+    type: 'dcc-legacy',
     name: 'DCC Registry',
     url: 'https://digitalcredentials.github.io/dcc-registry/registry.json'
   }